+++ ultimate/tests/org/jetbrains/kotlin/idea/spring/tests/inspections/AbstractSpringInspectionTest.kt
@@ -1,0 +1,11 @@
+package org.jetbrains.kotlin.idea.spring.tests.inspections
+
+import com.intellij.psi.PsiFile
+import org.jetbrains.kotlin.idea.codeInsight.AbstractInspectionTest
+import org.jetbrains.kotlin.idea.spring.tests.configureSpringFileSetByDirective
+
+abstract class AbstractSpringInspectionTest : AbstractInspectionTest() {
+    override fun configExtra(psiFiles: List<PsiFile>, options: String) {
+        configureSpringFileSetByDirective(myFixture.module, options, psiFiles)
+    }
+}