+++ ultimate/tests/org/jetbrains/kotlin/idea/spring/tests/quickfixes/AbstractSpringQuickFixTest.kt
@@ -1,0 +1,10 @@
+package org.jetbrains.kotlin.idea.spring.tests.quickfixes
+
+import org.jetbrains.kotlin.idea.quickfix.AbstractQuickFixTest
+import org.jetbrains.kotlin.idea.spring.tests.configureSpringFileSetByDirective
+
+abstract class AbstractSpringQuickFixTest : AbstractQuickFixTest() {
+    override fun configExtra(options: String) {
+        configureSpringFileSetByDirective(getModule(), options, listOf(getFile()))
+    }
+}